Update README.md
The Sakila database is a nicely normalised schema modelling a DVD rental store, featuring things like films, actors, film-actor relationships, and a central inventory table that connects films, stores, and rentals. This project uses SQL to analyze Sakila Dataset for insights and interesting data such as sakila's rental shop with the highest sales, List of the top 10 movies with the highest rental sales, The percentage of movie ratings (%) and else. This project is part of the Practice SQL EBOOK by DataRockie.